Move tabs into seperate files
